move bot config out of api
