+++ src/marvin/api/slackbot.py
@@ -1,15 +1,10 @@
 import asyncio
-import os
 
 import httpx
 from fastapi import Request, status
-from prefect.utilities.collections import listrepr
 from pydantic import BaseModel
 
-from marvin.bots import Bot
-from marvin.plugins.chroma import chroma_search
-from marvin.plugins.duckduckgo import DuckDuckGo
-from marvin.plugins.github import search_github_issues
+import marvin
 from marvin.utilities.types import MarvinRouter
 
 router = MarvinRouter(
@@ -17,73 +12,6 @@ router = MarvinRouter(
     prefix="/slack",
 )
 
-prefect_keywords = [
-    "prefect",
-    "cloud",
-    "server",
-    "ui",
-    "agent",
-    "flow",
-    "task",
-    "schedule",
-    "deployment",
-    "kubernetes",
-    "docker",
-    "aws",
-    "gcp",
-    "azure",
-    "ecs",
-    "fargate",
-    "lambda",
-    "s3",
-    "cloudwatch",
-    "dask",
-    "worker",
-    "work pool",
-    "k8s",
-    "helm",
-]
-
-how_to_write_a_prefect_2_flow = """
-from prefect import flow, task
-
-# This is a task decorator
-# this task will inherit the `log_prints` setting from the flow
-@task
-def say_hello(name: str):
-    print(f"Hello {name}!")
-
-# This is a flow decorator
-# it calls the `say_hello` task 3 times
-@flow(log_prints=True)
-def hello(name: str = "world", count: int = 1):
-    say_hello.map(f"{name}-{i}" for i in range(count))
-
-if __name__ == "__main__":
-    hello(count=3)
-"""
-
-chroma_search_instructions = (
-    "Do NOT answer ANY questions about Prefect without using one of the plugins. Use"
-    " the keywords provided below to decide whether a question is about Prefect. In"
-    " case you need it, here's an example of a Prefect 2 flow:"
-    f" {how_to_write_a_prefect_2_flow}. Use `chroma_search` to retrieve context when"
-    " asked a question containing any of the following keywords:"
-    f" {listrepr(prefect_keywords)}. If asked about a github issue, use the"
-    " `search_github_issues` plugin, choosing the most appropriate repo based on the"
-    " user's question. Always provide relevant links from plugin outputs. As a last"
-    " resort, use the `DuckDuckGo` plugin to search the web for answers to questions. "
-)
-
-community_bot = Bot(
-    name="Marvin",
-    personality="like the robot from HHGTTG, mildly depressed but helpful",
-    instructions=chroma_search_instructions,
-    plugins=[chroma_search, search_github_issues, DuckDuckGo()],
-)
-
-
-SLACK_BOT_TOKEN = os.environ.get("SLACK_BOT_TOKEN")
 SLACK_API_BASE_URL = "https://slack.com/api"
 
 
@@ -106,16 +34,23 @@ async def _post_message_to_slack(channel: str, message: str, thread_ts: str = No
 
         response = await client.post(
             f"{SLACK_API_BASE_URL}/chat.postMessage",
-            headers={"Authorization": f"Bearer {SLACK_BOT_TOKEN}"},
+            headers={
+                "Authorization": (
+                    "Bearer"
+                    f" {marvin.config.settings.slack_bot_token.get_secret_value()}"
+                )
+            },
             json=payload,
         )
         response.raise_for_status()
 
 
 async def _slack_response(event: SlackEvent):
-    await community_bot.set_thread(thread_lookup_key=f"{event.channel}:{event.user}")
+    bot = marvin.config.settings.slackbot
 
-    response = await community_bot.say(event.text)
+    await bot.set_thread(thread_lookup_key=f"{event.channel}:{event.user}")
+
+    response = await bot.say(event.text)
 
     await _post_message_to_slack(event.channel, response.content, event.ts)
 
@@ -127,7 +62,7 @@ async def handle_slack_events(request: Request):
     if payload["type"] == "url_verification":
         return payload["challenge"]
 
-    event = SlackEvent(**(await request.json())["event"])
+    event = SlackEvent(**payload["event"])
 
     if event.type == "app_mention":
         asyncio.ensure_future(_slack_response(event))
